requirements.txt
python-dotenv
requests
reportlab
sqlalchemy
pytz
